API/storages/es: refactoring.
One big file `storages/es.py` is turned into submodule:
* `__init__.py` for everything that might be needed from the outside;
* `methods.py` for methods implementations;
* `transform.py` for functions that transform ES response data into
  desired formats;
* `common.py` for common variables and functions for interaction with
  ES: query constructors and "get this or that data" functions.
