+++ Utils/API/server/lib/dkb/api/storages/es/__init__.py
@@ -1,0 +1,13 @@
+"""
+Interaction with DKB ES storage.
+"""
+
+from common import (STORAGE_NAME,
+                    QUERY_DIR)
+
+from methods import (task_steps_hist,
+                     task_chain,
+                     task_kwsearch,
+                     task_derivation_statistics,
+                     campaign_stat,
+                     step_stat)
